Retrieve page ids and change image
Running with --update retrieves all the page ids from the category
specified by `CATEGORY`. These are stored in the `images` table in the
database.

Running with --new-image picks an id at random from the `images` table,
downloads the corresponding image, if it doesn't alreay exist, and sets
it as desktop image. So far, this only works with MATE.

Log files are saved to `logs` directory. Logging to stderr as well is
enabled by --print-log.