Don't show old images and marking images as favourites
Added command to mark images as favourites. These will be kept and
considered again when picking a new image. Added value for when an
image was last shown, to not pick non favourite images that has
already been shown. The files for non favourite images are also
removed when a new image is shown.

Added argument to --fetch for specifying the number of ids that should
be fetched. Changed how the check for if an id is already in the
database is made. Increased the number of ids to get in the API query
to 500.

Changed stdout logging to show info level, rather than debug. Debug is
still used for the log files.